feat(desktop): rename Security settings tab to Remote Workspaces and add relay toggle to local host settings
- Rename the Security settings tab, page header, and command palette entry
  to Remote Workspaces (route stays /settings/security); keep it findable
  by searching "security"
- Extract the expose-via-relay toggle (paywall gate, typed confirmation,
  optimistic update) into a shared ExposeViaRelaySection component
- Show the relay toggle on the local host's settings page
- Simplify the toggle description and link to the remote-workspaces docs
- Update all "Settings > Security" copy in automations to the new name
