+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/automations/components/CreateAutomationDialog/CreateAutomationDialog.tsx
@@ -69,7 +69,7 @@ export function CreateAutomationDialog({
 	const { localHostId, localHostIsOnline } = useWorkspaceHostOptions();
 	const targetHostId = hostId ?? localHostId;
 	// The local device is only dispatchable once relay access is enabled in
-	// Settings > Security; block creation until then. Offline remote hosts
+	// Settings > Remote Workspaces; block creation until then. Offline remote hosts
 	// only warn — they may reconnect by the time a run is scheduled.
 	const localRelayBlocked =
 		targetHostId === localHostId && localHostIsOnline === false;
