+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/automations/utils/hostOfflineError/hostOfflineError.ts
@@ -4,7 +4,7 @@
 const HOST_OFFLINE_ERRORS = ["target host offline", "no host available"];
 
 export const HOST_OFFLINE_HELP =
-	"The host isn't connected to the Superset relay. If it's this device, turn on \"Allow remote workspaces to access this device via relay\" in Settings > Security, then try again.";
+	"The host isn't connected to the Superset relay. If it's this device, turn on \"Allow remote workspaces to access this device via relay\" in Settings > Remote Workspaces, then try again.";
 
 export function isHostOfflineError(error: string | null | undefined): boolean {
 	return !!error && HOST_OFFLINE_ERRORS.some((e) => error.includes(e));
