+++ apps/desktop/src/renderer/routes/_authenticated/settings/components/ExposeViaRelaySection/index.ts
@@ -1,0 +1,1 @@
+export { ExposeViaRelaySection } from "./ExposeViaRelaySection";
